flask-mysql/requirements.txt

Flask==3.1.1
Flask_JWT_Extended==4.7.1
mysql-connector-python==9.1.0
Werkzeug==3.1.3
